requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
python-ldap==3.4.3
PyMySQL==1.1.0
python-docx==0.8.11
pdfkit==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
Flask-WTF==1.2.1
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.2
bleach==6.1.0
